Do not crash when there is no diff
There might be a change without a diff, for instance when a file is replaced with its original version. We can think about what to do in such cases, but for now this patch would prevent the beacon from crashing. A restart of the minion seems to be necessary to recover from this crash.